new MailBoxReader class
the MailBoxReader class should make debugging easier in the future. we
can now interact with the MailBoxReader class directly by the following:

m = MailBoxReader()

All code related to the parsing of emails has now been integrated into
this class. There may be an opportunity to make this code base a bit
more effecient with a more straightforward approach.
